# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de pedidos de compra para manutenção de elevadores.",
        "title": "Elevator Purchase API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte",
            "email": "suporte@elevator.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/pedidos": {
            "get": {
                "description": "Lista todos os pedidos cadastrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Listar Pedidos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pedido.Pedido"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo pedido no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Cria um novo pedido",
                "parameters": [
                    {
                        "description": "Dados do Pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pedido.Pedido"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pedido.Pedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pedidos/{id}": {
            "put": {
                "description": "Atualiza os dados de um pedido existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Atualiza um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do Pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pedido.Pedido"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pedido.Pedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta um pedido pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Deleta um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido deletado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "pedido.Pedido": {
            "type": "object",
            "properties": {
                "cliente": {
                    "type": "string"
                },
                "dataAprovacao": {
                    "type": "string"
                },
                "dataEntrega": {
                    "type": "string"
                },
                "dataSolicitacao": {
                    "description": "Documentos      pq.StringArray `gorm:\"type:text[]\" json:\"documentos\"`",
                    "type": "string"
                },
                "documentos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fornecedor": {
                    "type": "string"
                },
                "foto": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "localArmazenamento": {
                    "type": "string"
                },
                "nomePeca": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/pedido.Status"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "pedido.Status": {
            "type": "string",
            "enum": [
                "cotado",
                "fazendo_orcamento",
                "orcamento",
                "compra_aprovada",
                "comprado"
            ],
            "x-enum-varnames": [
                "Cotado",
                "FazendoOrcamento",
                "Orcamento",
                "CompraAprovada",
                "Comprado"
            ]
        }
    }
}